chore: apply review feedback — trim comments, split transitioners per file
