+++ android/src/main/java/com/primerioreactnative/components/manager/NativeViewContainer.kt
@@ -8,9 +8,7 @@ import com.facebook.react.uimanager.ThemedReactContext
 
 class NativeViewContainer(private val context: ThemedReactContext) : FrameLayout(context) {
     fun addViewImpl(view: View) {
-        // The Klarna payment view is a single shared instance. If a new container is created (re-entry /
-        // re-render / Fabric view preallocation) while the previous one still holds it, addView throws
-        // "child already has a parent" and crashes the surface. Detach from the old parent first.
+        // Shared native view: detach from any previous container or addView throws "child already has a parent".
         (view.parent as? ViewGroup)?.removeView(view)
         addView(view)
         view.setOnClickListener {
